+++ tutorial/02_mario/go/pkg/define/define.go
@@ -1,8 +1,8 @@
 package define
 
 var (
-	WinWidth        = float32(1080)
-	WinHeight       = float32(1920)
+	WinWidth        = float64(1080)
+	WinHeight       = float64(1920)
 	Score     int64 = 0
 )
